+++ apps/web/src/useCases/pages/createPage/CreatePageForm.tsx
@@ -1,10 +1,5 @@
 import React from 'react'
-import {
-  ApolloForm,
-  ApolloFormProps,
-  JsonSchemaUseCaseFormProps,
-  PropsWithIds,
-} from '@codelab/frontend'
+import { ApolloForm, FormUseCaseProps, PropsWithIds } from '@codelab/frontend'
 import {
   CreatePageInputSchema,
   CreatePageMutationVariables,
@@ -13,9 +8,8 @@ import {
 } from '@codelab/generated'
 import { CreatePageInput } from 'libs/modules/page/src/core/application/useCases/createPage/CreatePageInput'
 
-type CreatePageFormProps = JsonSchemaUseCaseFormProps<CreatePageInput> &
-  PropsWithIds<'appId'> &
-  Pick<ApolloFormProps<any, any>, 'onSubmitSuccess'>
+type CreatePageFormProps = FormUseCaseProps<CreatePageInput> &
+  PropsWithIds<'appId'>
 
 export const CreatePageForm = ({ appId, ...props }: CreatePageFormProps) => {
   const [mutate] = useCreatePageMutation({
@@ -33,6 +27,7 @@ export const CreatePageForm = ({ appId, ...props }: CreatePageFormProps) => {
 
   return (
     <ApolloForm<CreatePageInput, CreatePageMutationVariables>
+      {...props}
       mutate={mutate}
       hideSubmitButton
       schema={CreatePageInputSchema}
@@ -44,7 +39,6 @@ export const CreatePageForm = ({ appId, ...props }: CreatePageFormProps) => {
         },
       }}
       initialFormData={{ title: '', appId }}
-      {...props}
     />
   )
 }
