+++ libs/frontend/src/components/form/json-schema/Form-jsonSchema.tsx
@@ -2,13 +2,11 @@ import { Theme as AntDTheme } from '@rjsf/antd'
 import { withTheme } from '@rjsf/core'
 import { Button } from 'antd'
 import React, { ReactElement, useState } from 'react'
-import { setSubmitControllerRef } from './JsonSchemaForm-ref'
-import { JsonSchemaFormProps } from './JsonSchemaForm.d'
+import { setSubmitControllerRef } from './Form-jsonSchema--ref'
+import { JsonSchemaFormProps } from './Form-jsonSchema--types'
 import { callCallbackOrArrayOfCallbacks } from 'libs/frontend/src/utils'
 
 const ThemedForm = withTheme(AntDTheme)
-
-/** Generic form event containing the current form values  */
 
 /**
  * A form that's generated by a JSON schema.
