+++ libs/frontend/src/components/form/json-schema/index.ts
@@ -1,2 +1,3 @@
-export * from './JsonSchemaForm'
-export * from './JsonSchemaForm.d'
+export * from './Form-jsonSchema'
+export * from './Form-jsonSchema--types'
+export * from './Form-jsonSchema--ref'
