+++ libs/frontend/src/components/index.ts
@@ -1,5 +1,4 @@
-export * from './form/ModalForm'
+export * from './form/Form-modal'
 export * from './form/json-schema'
-export * from './form/ApolloForm'
-export * from './form/ApolloForm.d'
-export * from './form/XStateForm'
+export * from './form/apollo/Form-apollo'
+export * from './form/apollo/Form-apollo--types'
